refactor state upgrades
