+++ params/extras/state_upgrade.go
@@ -5,63 +5,16 @@ package extras
 
 import (
 	"fmt"
-	"reflect"
 
-	"github.com/ava-labs/libevm/common"
-	"github.com/ava-labs/libevm/common/hexutil"
-	"github.com/ava-labs/libevm/common/math"
+	"github.com/ava-labs/subnet-evm/stateupgrade"
 
 	ethparams "github.com/ava-labs/libevm/params"
 )
 
-// StateUpgrade describes the modifications to be made to the state during
-// a state upgrade.
-type StateUpgrade struct {
-	BlockTimestamp *uint64 `json:"blockTimestamp,omitempty"`
-
-	// map from account address to the modification to be made to the account.
-	StateUpgradeAccounts map[common.Address]StateUpgradeAccount `json:"accounts"`
-}
-
-// StateUpgradeAccount describes the modifications to be made to an account during
-// a state upgrade.
-type StateUpgradeAccount struct {
-	Code          hexutil.Bytes               `json:"code,omitempty"`
-	Storage       map[common.Hash]common.Hash `json:"storage,omitempty"`
-	BalanceChange *math.HexOrDecimal256       `json:"balanceChange,omitempty"`
-}
-
-func (s *StateUpgrade) Equal(other *StateUpgrade) bool {
-	return reflect.DeepEqual(s, other)
-}
-
-// verifyStateUpgrades checks [c.StateUpgrades] is well formed:
-// - the specified blockTimestamps must monotonically increase
-func (c *ChainConfig) verifyStateUpgrades() error {
-	var previousUpgradeTimestamp *uint64
-	for i, upgrade := range c.StateUpgrades {
-		upgradeTimestamp := upgrade.BlockTimestamp
-		if upgradeTimestamp == nil {
-			return fmt.Errorf("StateUpgrade[%d]: config block timestamp cannot be nil ", i)
-		}
-		// Verify the upgrade's timestamp is equal 0 (to avoid confusion with genesis).
-		if *upgradeTimestamp == 0 {
-			return fmt.Errorf("StateUpgrade[%d]: config block timestamp (%v) must be greater than 0", i, *upgradeTimestamp)
-		}
-
-		// Verify specified timestamps are strictly monotonically increasing.
-		if previousUpgradeTimestamp != nil && *upgradeTimestamp <= *previousUpgradeTimestamp {
-			return fmt.Errorf("StateUpgrade[%d]: config block timestamp (%v) <= previous timestamp (%v)", i, *upgradeTimestamp, *previousUpgradeTimestamp)
-		}
-		previousUpgradeTimestamp = upgradeTimestamp
-	}
-	return nil
-}
-
 // GetActivatingStateUpgrades returns all state upgrades configured to activate during the
 // state transition from a block with timestamp [from] to a block with timestamp [to].
-func (*ChainConfig) GetActivatingStateUpgrades(from *uint64, to uint64, upgrades []StateUpgrade) []StateUpgrade {
-	activating := make([]StateUpgrade, 0)
+func (*ChainConfig) GetActivatingStateUpgrades(from *uint64, to uint64, upgrades []stateupgrade.StateUpgrade) []stateupgrade.StateUpgrade {
+	activating := make([]stateupgrade.StateUpgrade, 0)
 	for _, upgrade := range upgrades {
 		if IsForkTransition(upgrade.BlockTimestamp, from, to) {
 			activating = append(activating, upgrade)
@@ -71,7 +24,7 @@ func (*ChainConfig) GetActivatingStateUpgrades(from *uint64, to uint64, upgrades
 }
 
 // checkStateUpgradesCompatible checks if [stateUpgrades] are compatible with [c] at [headTimestamp].
-func (c *ChainConfig) checkStateUpgradesCompatible(stateUpgrades []StateUpgrade, lastTimestamp uint64) *ethparams.ConfigCompatError {
+func (c *ChainConfig) checkStateUpgradesCompatible(stateUpgrades []stateupgrade.StateUpgrade, lastTimestamp uint64) *ethparams.ConfigCompatError {
 	// All active upgrades (from nil to [lastTimestamp]) must match.
 	activeUpgrades := c.GetActivatingStateUpgrades(nil, lastTimestamp, c.StateUpgrades)
 	newUpgrades := c.GetActivatingStateUpgrades(nil, lastTimestamp, stateUpgrades)
