+++ params/extras/state_upgrade_test.go
@@ -12,76 +12,16 @@ import (
 	"github.com/ava-labs/libevm/common/math"
 	"github.com/stretchr/testify/require"
 
+	"github.com/ava-labs/subnet-evm/stateupgrade"
 	"github.com/ava-labs/subnet-evm/utils"
-	"github.com/ava-labs/subnet-evm/utils/utilstest"
 )
-
-func TestVerifyStateUpgrades(t *testing.T) {
-	modifiedAccounts := map[common.Address]StateUpgradeAccount{
-		{1}: {
-			BalanceChange: (*math.HexOrDecimal256)(common.Big1),
-		},
-	}
-	tests := []struct {
-		name          string
-		upgrades      []StateUpgrade
-		expectedError string
-	}{
-		{
-			name: "valid upgrade",
-			upgrades: []StateUpgrade{
-				{BlockTimestamp: utils.NewUint64(1), StateUpgradeAccounts: modifiedAccounts},
-				{BlockTimestamp: utils.NewUint64(2), StateUpgradeAccounts: modifiedAccounts},
-			},
-		},
-		{
-			name: "upgrade block timestamp is not strictly increasing",
-			upgrades: []StateUpgrade{
-				{BlockTimestamp: utils.NewUint64(1), StateUpgradeAccounts: modifiedAccounts},
-				{BlockTimestamp: utils.NewUint64(1), StateUpgradeAccounts: modifiedAccounts},
-			},
-			expectedError: "config block timestamp (1) <= previous timestamp (1)",
-		},
-		{
-			name: "upgrade block timestamp decreases",
-			upgrades: []StateUpgrade{
-				{BlockTimestamp: utils.NewUint64(2), StateUpgradeAccounts: modifiedAccounts},
-				{BlockTimestamp: utils.NewUint64(1), StateUpgradeAccounts: modifiedAccounts},
-			},
-			expectedError: "config block timestamp (1) <= previous timestamp (2)",
-		},
-		{
-			name: "upgrade block timestamp is zero",
-			upgrades: []StateUpgrade{
-				{BlockTimestamp: utils.NewUint64(0), StateUpgradeAccounts: modifiedAccounts},
-			},
-			expectedError: "config block timestamp (0) must be greater than 0",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			require := require.New(t)
-			c := *TestChainConfig
-			config := &c
-			config.SnowCtx = utilstest.NewTestSnowContext(t)
-			config.StateUpgrades = tt.upgrades
-
-			err := config.Verify()
-			if tt.expectedError == "" {
-				require.NoError(err)
-			} else {
-				require.ErrorContains(err, tt.expectedError)
-			}
-		})
-	}
-}
 
 func TestCheckCompatibleStateUpgrades(t *testing.T) {
 	chainConfig := *TestChainConfig
-	stateUpgrade := map[common.Address]StateUpgradeAccount{
+	stateUpgrade := map[common.Address]stateupgrade.StateUpgradeAccount{
 		{1}: {BalanceChange: (*math.HexOrDecimal256)(common.Big1)},
 	}
-	differentStateUpgrade := map[common.Address]StateUpgradeAccount{
+	differentStateUpgrade := map[common.Address]stateupgrade.StateUpgradeAccount{
 		{2}: {BalanceChange: (*math.HexOrDecimal256)(common.Big1)},
 	}
 
@@ -90,12 +30,12 @@ func TestCheckCompatibleStateUpgrades(t *testing.T) {
 			startTimestamps: []uint64{5, 6},
 			configs: []*UpgradeConfig{
 				{
-					StateUpgrades: []StateUpgrade{
+					StateUpgrades: []stateupgrade.StateUpgrade{
 						{BlockTimestamp: utils.NewUint64(6), StateUpgradeAccounts: stateUpgrade},
 					},
 				},
 				{
-					StateUpgrades: []StateUpgrade{
+					StateUpgrades: []stateupgrade.StateUpgrade{
 						{BlockTimestamp: utils.NewUint64(6), StateUpgradeAccounts: stateUpgrade},
 					},
 				},
@@ -106,13 +46,13 @@ func TestCheckCompatibleStateUpgrades(t *testing.T) {
 			startTimestamps:     []uint64{5, 8},
 			configs: []*UpgradeConfig{
 				{
-					StateUpgrades: []StateUpgrade{
+					StateUpgrades: []stateupgrade.StateUpgrade{
 						{BlockTimestamp: utils.NewUint64(6), StateUpgradeAccounts: stateUpgrade},
 						{BlockTimestamp: utils.NewUint64(7), StateUpgradeAccounts: stateUpgrade},
 					},
 				},
 				{
-					StateUpgrades: []StateUpgrade{
+					StateUpgrades: []stateupgrade.StateUpgrade{
 						{BlockTimestamp: utils.NewUint64(6), StateUpgradeAccounts: stateUpgrade},
 						{BlockTimestamp: utils.NewUint64(7), StateUpgradeAccounts: differentStateUpgrade},
 					},
@@ -123,13 +63,13 @@ func TestCheckCompatibleStateUpgrades(t *testing.T) {
 			startTimestamps: []uint64{5, 6},
 			configs: []*UpgradeConfig{
 				{
-					StateUpgrades: []StateUpgrade{
+					StateUpgrades: []stateupgrade.StateUpgrade{
 						{BlockTimestamp: utils.NewUint64(6), StateUpgradeAccounts: stateUpgrade},
 						{BlockTimestamp: utils.NewUint64(7), StateUpgradeAccounts: stateUpgrade},
 					},
 				},
 				{
-					StateUpgrades: []StateUpgrade{
+					StateUpgrades: []stateupgrade.StateUpgrade{
 						{BlockTimestamp: utils.NewUint64(6), StateUpgradeAccounts: stateUpgrade},
 					},
 				},
@@ -140,7 +80,7 @@ func TestCheckCompatibleStateUpgrades(t *testing.T) {
 			startTimestamps:     []uint64{6},
 			configs: []*UpgradeConfig{
 				{
-					StateUpgrades: []StateUpgrade{
+					StateUpgrades: []stateupgrade.StateUpgrade{
 						{BlockTimestamp: utils.NewUint64(5), StateUpgradeAccounts: stateUpgrade},
 					},
 				},
@@ -172,10 +112,10 @@ func TestUnmarshalStateUpgradeJSON(t *testing.T) {
 	)
 
 	upgradeConfig := UpgradeConfig{
-		StateUpgrades: []StateUpgrade{
+		StateUpgrades: []stateupgrade.StateUpgrade{
 			{
 				BlockTimestamp: utils.NewUint64(1677608400),
-				StateUpgradeAccounts: map[common.Address]StateUpgradeAccount{
+				StateUpgradeAccounts: map[common.Address]stateupgrade.StateUpgradeAccount{
 					common.HexToAddress("0x8db97C7cEcE249c2b98bDC0226Cc4C2A57BF52FC"): {
 						BalanceChange: (*math.HexOrDecimal256)(big.NewInt(100)),
 					},
